requirements.txt
# PromptToProduct - Agent Orchestration System Dependencies

# Core dependencies (currently in use)
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0

# LangGraph dependencies for advanced agent workflows
langgraph>=0.0.32
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Additional LangGraph ecosystem packages
langsmith>=0.0.83

# Optional but recommended for enhanced functionality
python-multipart>=0.0.6
uvicorn>=0.24.0

# Playwright MCP Server for web automation
playwright-mcp>=0.1.0
playwright>=1.50.0
mcp>=1.3.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Documentation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.4.0